requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
--extra-index-url https://lambda_user:VP8lvTPNBd@pypi.syse-i.net/simple
appnope==0.1.2; sys_platform == 'darwin'
asgiref==3.3.1; python_version >= '3.5'
backcall==0.2.0
bcrypt==3.2.0; python_version >= '3.6'
boto3==1.16.60
botocore==1.19.60
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==4.4.2
defusedxml==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.44.0
django-compat==1.0.15
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-guardian==2.3.0
django-hijack==2.1.10
django-lambda-config==0.0.7
django-lambda-theme-tailwindcss==0.1.2+ga759031.dirty
django-lambda-theme==0.35.4
django-lambda-url==0.0.6
django-maintenance-mode==0.15.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-rest-knox==4.1.0
django-reversion==3.0.9
django-sendgrid-v5==0.9.0
django-storages==1.11.1
django==3.1.5
djangorestframework==3.12.2
fabric==2.6.0
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
invoke==1.5.0
ipython-genutils==0.2.0
ipython==7.19.0
jedi==0.18.0; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonfield==3.1.0; python_version >= '3.6'
markdown==3.3.3
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
paramiko==2.7.2
parso==0.8.1; python_version >= '3.6'
pathlib2==2.3.5
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==8.1.0
prompt-toolkit==3.0.14; python_full_version >= '3.6.1'
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.7.4; python_version >= '3.5'
pyjwt[crypto]==2.0.1; python_version >= '3.6'
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-http-client==3.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2020.5
pyyaml==5.4.1
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
s3transfer==0.3.4
sendgrid==6.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
sentry-sdk==0.19.5
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
starkbank-ecdsa==1.1.0
swapper==1.1.2.post1
traitlets==5.0.5; python_version >= '3.7'
urllib3==1.26.3; python_version != '3.4'
wcwidth==0.2.5
